todoqueue_backend/requirements.txt
asgiref==3.7.2
distlib==0.3.5
Django==4.2.5
django-cors-headers==4.2.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
filelock==3.8.0
gunicorn==21.2.0
mysqlclient==2.2.0
packaging==23.1
platformdirs==2.5.2
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.7.1
virtualenv==20.16.3
